# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "user_findMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "user_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "user_deleteMe",
        "parameters": [],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/taste": {
      "get": {
        "operationId": "user_getMyTaste",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TasteStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/rating-scale": {
      "get": {
        "operationId": "user_getRatingScale",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingScaleRuleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Какую шкалу оценок показывать для каждого типа тайтла",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/password": {
      "patch": {
        "operationId": "user_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/public/{username}": {
      "get": {
        "operationId": "user_findPublic",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfileResponse"
                }
              }
            }
          }
        },
        "summary": "Публичный Taste DNA по ссылке — открывается без авторизации",
        "tags": [
          "users"
        ]
      }
    },
    "/titles": {
      "get": {
        "operationId": "title_findAll",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Несколько типов сразу: ?type=MOVIE&type=GAME",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "GAME",
                  "MOVIE",
                  "TV_SHOW",
                  "BOOK",
                  "ANIME"
                ]
              }
            }
          },
          {
            "name": "genres",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "minRating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 10,
              "type": "number"
            }
          },
          {
            "name": "yearFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "yearTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "popular",
              "type": "string",
              "enum": [
                "name",
                "rating",
                "newest",
                "popular"
              ]
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TitleListResponse"
                }
              }
            }
          }
        },
        "tags": [
          "titles"
        ]
      }
    },
    "/titles/{slug}": {
      "get": {
        "operationId": "title_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TitleResponse"
                }
              }
            }
          }
        },
        "tags": [
          "titles"
        ]
      }
    },
    "/search": {
      "get": {
        "operationId": "search_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 2,
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "GAME",
                  "MOVIE",
                  "TV_SHOW",
                  "BOOK",
                  "ANIME"
                ]
              }
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TitleListResponse"
                }
              }
            }
          }
        },
        "tags": [
          "search"
        ]
      }
    },
    "/search/suggest": {
      "get": {
        "operationId": "search_suggest",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TitleListItemResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "Подсказки для поисковой строки и браузерного расширения",
        "tags": [
          "search"
        ]
      }
    },
...
    "/subscription": {
      "get": {
        "operationId": "subscription_findMy",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Тариф, лимиты и текущее потребление",
        "tags": [
          "subscription"
        ]
      }
    },
    "/subscription/verify": {
      "post": {
        "operationId": "subscription_verifyPurchase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPurchaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Подтверждение покупки из мобильного приложения.\nВеб и расширение сюда не ходят: оплата только в сторах",
        "tags": [
          "subscription"
        ]
      }
    },
    "/subscription/trial": {
      "post": {
        "operationId": "subscription_startTrial",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subscription"
        ]
      }
    },
    "/subscription/cancel": {
      "post": {
        "operationId": "subscription_cancel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subscription"
        ]
      }
    },
...
    "/reviews/title/{slug}": {
      "get": {
        "operationId": "review_findByTitle",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "newest",
              "type": "string",
              "enum": [
                "rating",
                "newest",
                "ratingAsc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            }
          }
        },
        "summary": "Отзывы на тайтле — открыты без авторизации",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "RED Marathōn API",
    "description": "One library for movies, TV shows, anime, books and games",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "ratingScale": {
            "type": "string",
            "description": "Как показывать оценки. Хранятся всегда по 10-балльной шкале.\nAUTO — шкала выбирается по типу тайтла, см. /users/me/rating-scales",
            "enum": [
              "AUTO",
              "FIVE",
              "TEN"
            ]
          }
        },
        "required": [
          "displayName",
          "bio",
          "avatarUrl",
          "country",
          "birthDate",
          "isPublic",
          "ratingScale"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "enum": [
              "USER",
              "MODERATOR",
              "ADMIN"
            ],
            "type": "string"
          },
          "isVerified": {
            "type": "boolean"
          },
          "profile": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProfileResponse"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "role",
          "isVerified",
          "profile",
          "createdAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 50
          },
          "bio": {
            "type": "string",
            "maxLength": 500
          },
          "avatarUrl": {
            "type": "string",
            "maxLength": 500
          },
          "country": {
            "type": "string",
            "maxLength": 60
          },
          "birthDate": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean",
            "description": "Открыт ли Taste DNA по публичной ссылке"
          },
          "ratingScale": {
            "type": "string",
            "description": "Предпочтительная шкала отображения. На бэкенде оценки всегда\nдесятибалльные — фронт делит на 2, когда выбрана пятибалльная",
            "enum": [
              "AUTO",
              "FIVE",
              "TEN"
            ]
          }
        }
      },
      "TasteStatsResponse": {
        "type": "object",
        "properties": {
          "byType": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            },
            "description": "Сколько тайтлов в библиотеке по каждому типу"
          },
          "byStatus": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            },
            "description": "Сколько записей в каждом статусе"
          },
          "totalEntries": {
            "type": "number"
          },
          "completedCount": {
            "type": "number"
          },
          "averageRating": {
            "type": "number",
            "nullable": true
          },
          "favoriteGenres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "byType",
          "byStatus",
          "totalEntries",
          "completedCount",
          "averageRating",
          "favoriteGenres"
        ]
      },
      "RatingScaleRuleResponse": {
        "type": "object",
        "properties": {
          "scale": {
            "type": "string",
            "description": "Что выбрал пользователь",
            "enum": [
              "AUTO",
              "FIVE",
              "TEN"
            ]
          },
          "byType": {
            "type": "object",
            "additionalProperties": true,
            "description": "Какую шкалу показывать для каждого типа при текущей настройке"
          }
        },
        "required": [
          "scale",
          "byType"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "minLength": 6,
            "maxLength": 100
          },
          "newPassword": {
            "type": "string",
            "minLength": 6,
            "maxLength": 100
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "PublicProfileResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "taste": {
            "$ref": "#/components/schemas/TasteStatsResponse"
          }
        },
        "required": [
          "username",
          "displayName",
          "bio",
          "avatarUrl",
          "country",
          "createdAt",
          "taste"
        ]
      },
      "TitleListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "GAME",
              "MOVIE",
              "TV_SHOW",
              "BOOK",
              "ANIME"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number"
          },
          "ratingCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "slug",
          "coverUrl",
          "releaseDate",
          "rating",
          "ratingCount"
        ]
      },
      "TitleListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TitleListItemResponse"
            }
          },
          "isHasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "isHasMore"
        ]
      },
      "GenreResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "ActorResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "photoUrl"
        ]
      },
      "TitleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "GAME",
              "MOVIE",
              "TV_SHOW",
              "BOOK",
              "ANIME"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number"
          },
          "ratingCount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED"
            ]
          },
          "originalName": {
            "type": "string",
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreResponse"
            }
          },
          "similar": {
            "description": "Похожие тайтлы — приходят вместе со страницей, отдельный запрос не нужен",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TitleListItemResponse"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Ниже — из внешнего API, в нашей базе не хранится"
          },
          "actors": {
            "description": "Актёры у фильмов, студии у игр и аниме, авторы у книг",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActorResponse"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Страницы книги, платформы игры, число серий"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "slug",
          "coverUrl",
          "releaseDate",
          "rating",
          "ratingCount",
          "status",
          "originalName",
          "genres",
          "similar",
          "createdAt",
          "description",
          "actors",
          "metadata"
        ]
      },
...
      "SubscriptionLimitsResponse": {
        "type": "object",
        "properties": {
          "collections": {
            "type": "number",
            "nullable": true,
            "description": "null означает «без ограничений»"
          },
          "libraryEntries": {
            "type": "number",
            "nullable": true
          },
          "aiRequestsPerDay": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "collections",
          "libraryEntries",
          "aiRequestsPerDay"
        ]
      },
      "SubscriptionUsageResponse": {
        "type": "object",
        "properties": {
          "collections": {
            "type": "number"
          },
          "libraryEntries": {
            "type": "number"
          }
        },
        "required": [
          "collections",
          "libraryEntries"
        ]
      },
      "SubscriptionResponse": {
        "type": "object",
        "properties": {
          "plan": {
            "enum": [
              "FREE",
              "PRO"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "TRIALING",
              "PAST_DUE",
              "CANCELED",
              "EXPIRED"
            ],
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "endsAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "canceledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "autoRenew": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean",
            "description": "Активна ли подписка прямо сейчас"
          },
          "limits": {
            "$ref": "#/components/schemas/SubscriptionLimitsResponse"
          },
          "usage": {
            "$ref": "#/components/schemas/SubscriptionUsageResponse"
          }
        },
        "required": [
          "plan",
          "status",
          "startsAt",
          "endsAt",
          "canceledAt",
          "autoRenew",
          "isActive",
          "limits",
          "usage"
        ]
      },
      "VerifyPurchaseDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "APPLE",
              "GOOGLE"
            ]
          },
          "token": {
            "type": "string",
            "minLength": 4,
            "maxLength": 4000
          }
        },
        "required": [
          "provider",
          "token"
        ]
      },
...
      "ReviewListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            }
          },
          "isHasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "isHasMore"
        ]
      },
...
    }
  }
}